third_party/blink/web_tests/platform/generic/fast/forms/reportValidity-invalid-expected.txt

This test checks if reportValidity() returns correctly a false (meaning error) result on invalid elements, and returns a true result on a blank but valid elements. Blank but non-placeholder label options are valid.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

   
empty
another
 
X
 
X
PASS reportValidityFor("input-empty") is false
PASS document.activeElement is $("input-empty")
PASS reportValidityFor("input-pattern-mismatch") is false
PASS document.activeElement is $("input-pattern-mismatch")
PASS reportValidityFor("textarea") is false
PASS document.activeElement is $("textarea")
PASS reportValidityFor("select-no-explicit-value") is true
PASS document.activeElement is not $("select-no-explicit-value")
PASS reportValidityFor("select-placeholder") is false
PASS document.activeElement is $("select-placeholder")
PASS reportValidityFor("select-non-placeholder") is true
PASS document.activeElement is not $("select-non-placeholder")
PASS reportValidityFor("sad-form") is false
PASS document.activeElement is $("input-empty")
PASS successfullyParsed is true

TEST COMPLETE
